Fix the ZIP codes caching strategy over multiple processes
The scope manager persists the caching in one process but if
another process changes the state, this is not updated in the
original process that loaded the configuration before that.

This fixes the issue by using `Rails.cache` instead of local
memory cache.

The problem may still occur if the memory cache is used for
production configuration utilizing multiple processes.