Simplify the local caching of the user ZIP code
In order to fix some caching bugs, move the ZIP code local caching
to the workflow class and refer to the ZIP code through that class
with context information instead of fetching it through the
various scope managers.

Removes also permanent process-level caching of the user ZIP codes
and instead stores it locally once per request.